requirements.txt
blinker==1.9.0
boto3==1.42.44
botocore==1.42.44
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
phonenumbers==9.0.23
pika==1.3.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.5
